sr(api): move training-trigger-repository in devcomp
