+++ api/tests/devcomp/integration/infrastructure/repositories/training-trigger-repository_test.js
@@ -7,14 +7,14 @@ import {
   learningContentBuilder,
   mockLearningContent,
   sinon,
-} from '../../../test-helper.js';
+} from '../../../../test-helper.js';
 
-import * as trainingTriggerRepository from '../../../../lib/infrastructure/repositories/training-trigger-repository.js';
-import { TrainingTrigger, TrainingTriggerTube } from '../../../../lib/domain/models/index.js';
-import { TrainingTriggerForAdmin } from '../../../../lib/domain/read-models/TrainingTriggerForAdmin.js';
+import * as trainingTriggerRepository from '../../../../../src/devcomp/infrastructure/repositories/training-trigger-repository.js';
+import { TrainingTrigger, TrainingTriggerTube } from '../../../../../lib/domain/models/index.js';
+import { TrainingTriggerForAdmin } from '../../../../../lib/domain/read-models/TrainingTriggerForAdmin.js';
 import _ from 'lodash';
-import { NotFoundError } from '../../../../lib/domain/errors.js';
-import { logger } from '../../../../lib/infrastructure/logger.js';
+import { NotFoundError } from '../../../../../lib/domain/errors.js';
+import { logger } from '../../../../../lib/infrastructure/logger.js';
 
 describe('Integration | Repository | training-trigger-repository', function () {
   let learningContent;
